Restructure configuration package
