+++ carim/configuration/__init__.py
@@ -1,5 +1,5 @@
 def scan():
-    import carim.configuration.profiles
-    import carim.configuration.missions
-    import carim.configuration.auth
-    import carim.configuration.trader
+    import carim.configuration.universal
+    import carim.configuration.mods
+    import carim.configuration.omega
+    import carim.configuration.server
