+++ carim/configuration/mods/__init__.py
@@ -1,0 +1,11 @@
+import carim.configuration.mods.airdrop
+import carim.configuration.mods.cl0uds
+import carim.configuration.mods.code_lock
+import carim.configuration.mods.mass
+import carim.configuration.mods.msfc
+import carim.configuration.mods.munghards
+import carim.configuration.mods.server_panel
+import carim.configuration.mods.simple_base
+import carim.configuration.mods.trader
+import carim.configuration.mods.vanilla_plus_plus_map
+import carim.configuration.mods.vpp_admin_tools
