+++ carim/configuration/mods/trader/models/config.py
