+++ carim/configuration/server/missions.py
@@ -6,19 +6,14 @@ import re
 from xml.dom import minidom
 from xml.etree import ElementTree
 
-from carim.configuration import base
-from carim.models import types, deploydir, vpp_map
+from carim.configuration import decorators
+from carim.global_resources import types, deploydir, locations
 from carim.util import file_writing
 
 log = logging.getLogger(__name__)
 
 
-def mission(_func=None, *, directory='.', register=True):
-    return base.located_config(_func, directory=directory, dir_prefix='servers/0/mpmissions/dayzOffline.chernarusplus',
-                               register=register)
-
-
-@mission(directory='db', register=False)
+@decorators.mission(directory='db', register=False)
 def types_config(directory):
     types.get().getroot()[:] = sorted(types.get().getroot(), key=lambda child: child.get('name').lower())
     rough_string = ElementTree.tostring(types.get().getroot(), encoding='unicode')
@@ -29,7 +24,7 @@ def types_config(directory):
         f.write(reparsed.toprettyxml(indent='  ', newl=''))
 
 
-@mission(directory='db')
+@decorators.mission(directory='db')
 def globals_config(directory):
     globals_xml = ElementTree.parse(
         pathlib.Path(deploydir.get(), 'mpmissions/dayzOffline.chernarusplus/db/globals.xml'))
@@ -42,7 +37,7 @@ def globals_config(directory):
         f.write(file_writing.convert_to_string(globals_xml.getroot()))
 
 
-@mission(directory='db')
+@decorators.mission(directory='db')
 def events_config(directory):
     events_xml = ElementTree.parse(
         pathlib.Path(deploydir.get(), 'mpmissions/dayzOffline.chernarusplus/db/events.xml'))
@@ -60,13 +55,13 @@ def events_config(directory):
         f.write(file_writing.convert_to_string(events_xml.getroot()))
 
 
-@mission
+@decorators.mission
 def map_config(directory):
     file_writing.copy('resources/modifications/server/chernarusplus_tiers_have_traders_removed.map',
                       pathlib.Path(directory, 'areaflags.map'))
 
 
-@mission(directory='env')
+@decorators.mission(directory='env')
 def territory_config(directory):
     with open('resources/modifications/server/territories.json') as f:
         territories_modifications = json.load(f)
@@ -93,9 +88,9 @@ def remove_zones_if_near_traders(territory, name):
         x = float(raw[0])
         z = float(raw[1])
         r = float(raw[2])
-        clean_traders = (mark[1] for mark in vpp_map.marks[:3])
+        clean_traders = (mark[1] for mark in locations.marks[:3])
         for position in clean_traders:
-            if vpp_map.overlaps(position, 500, x, z, r):
+            if locations.overlaps(position, 500, x, z, r):
                 find_string = './/zone[@x="{}"][@z="{}"][@r="{}"]...'.format(*raw)
                 parents = territory.findall(find_string)
                 for parent in parents:
@@ -108,7 +103,7 @@ def remove_zones_if_near_traders(territory, name):
         log.info('removed {} zones from {}'.format(count, name))
 
 
-@mission
+@decorators.mission
 def remove_building_spawns_near_traders(directory):
     p = pathlib.Path(deploydir.get(), 'mpmissions/dayzOffline.chernarusplus/mapgrouppos.xml')
     count = 0
@@ -117,9 +112,9 @@ def remove_building_spawns_near_traders(directory):
         raw = group.get('pos')
         # log.info('{} {}'.format(group.get('name'), raw))
         x, y, z = (float(i) for i in raw.split(' '))
-        clean_traders = (mark[1] for mark in vpp_map.marks[:4])
+        clean_traders = (mark[1] for mark in locations.marks[:4])
         for position in clean_traders:
-            if vpp_map.overlaps(position, 200, x, z, 1):
+            if locations.overlaps(position, 200, x, z, 1):
                 find_string = './/group[@pos="{}"]...'.format(raw)
                 parents = mapgroups.findall(find_string)
                 for parent in parents:
