+++ carim/configuration/universal/__init__.py
@@ -1,0 +1,1 @@
+import carim.configuration.universal.auth
