+++ carim/global_resources/auth.py
@@ -1,10 +1,10 @@
 _auth = {}
 
 
+def get():
+    return _auth
+
+
 def set(auth):
     global _auth
     _auth = auth
-
-
-def get():
-    return _auth
