+++ carim/global_resources/configs.py
@@ -1,9 +1,9 @@
 _configs = []
 
 
-def add(config):
-    _configs.append(config)
-
-
 def get():
     return _configs
+
+
+def add(config):
+    _configs.append(config)
