+++ carim/global_resources/errors.py
@@ -1,9 +1,9 @@
 _errors = []
 
 
-def add(e):
-    _errors.append(e)
-
-
 def get():
     return _errors
+
+
+def add(e):
+    _errors.append(e)
