+++ carim/global_resources/item_type.py
