+++ carim/global_resources/outdir.py
