+++ carim/global_resources/types.py
@@ -1,10 +1,10 @@
 _types = None
 
 
+def get():
+    return _types
+
+
 def set(types):
     global _types
     _types = types
-
-
-def get():
-    return _types
