+++ carim/util/modify_types.py
@@ -3,13 +3,13 @@ import logging
 import re
 from xml.etree import ElementTree
 
-from carim.configuration import base
-from carim.models import types, item_type
+from carim.configuration import decorators
+from carim.global_resources import types, item_type
 
 log = logging.getLogger(__name__)
 
 
-@base.server(directory='mpmissions/dayzOffline.chernarusplus/db', register=False)
+@decorators.server(directory='mpmissions/dayzOffline.chernarusplus/db', register=False)
 def modify_types(directory):
     with open('resources/modifications/server/types_config.json') as f:
         type_config = json.load(f)
